std: update pipe tests
